Add phase gap backend check
