+++ client/src/locales/en.json
@@ -176,7 +176,8 @@
   "signupError": {
     "generic": "Signup failed",
     "signupEnded": "Signup ended",
-    "samePriority": "Priority already selected"
+    "samePriority": "Priority already selected",
+    "phaseGap": "Signup not open yet"
   },
   "signupUnsavedChanges": "Unsaved changes",
   "signupResultHint": "Sign up results are visible on 'My Games' page at",
